+++ tests/FunctionTest.php
@@ -17,9 +17,11 @@ use DateTime;
 use DateTimeImmutable;
 use DateTimeInterface;
 use Exception as PhpException;
+use League\Period\Duration;
 use League\Period\Exception;
 use PHPUnit\Framework\TestCase;
 use TypeError;
+use function get_object_vars;
 use function League\Period\datepoint;
 use function League\Period\day;
 use function League\Period\duration;
@@ -83,11 +85,12 @@ class FunctionTest extends TestCase
     /**
      * @dataProvider durationProvider
      *
+     * @param mixed                   $expected DateInterval object
      * @param int|DateInterval|string $duration
      */
-    public function testDuration(DateInterval $expected, $duration): void
+    public function testDuration($expected, $duration): void
     {
-        self::assertEquals($expected, duration($duration));
+        self::assertEquals(get_object_vars($expected), get_object_vars(duration($duration)));
     }
 
     public function durationProvider(): array
@@ -98,7 +101,7 @@ class FunctionTest extends TestCase
                 'input' => new DateInterval('P1D'),
             ],
             'string' => [
-                'expected' => new DateInterval('P1D'),
+                'expected' => new Duration('P1D'),
                 'input' => '+1 DAY',
             ],
             'int' => [
@@ -328,8 +331,6 @@ class FunctionTest extends TestCase
         self::assertEquals(new DateTimeImmutable('2008-07-02T00:00:00+08:00'), $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
     }
 
     public function testAlternateDay(): void
@@ -353,8 +354,6 @@ class FunctionTest extends TestCase
         self::assertEquals(new DateTimeImmutable('2008-07-01T23:00:00+08:00'), $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
     }
 
     public function testAlternateHour(): void
@@ -380,8 +379,6 @@ class FunctionTest extends TestCase
         self::assertEquals(new DateTimeImmutable('2008-07-01T22:36:00+08:00'), $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
     }
 
     public function testAlternateMinute(): void
@@ -410,8 +407,6 @@ class FunctionTest extends TestCase
         self::assertEquals(new DateTimeImmutable('2008-07-01T22:35:18+08:00'), $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
     }
 
     public function testAlternateSecond(): void
@@ -439,8 +434,6 @@ class FunctionTest extends TestCase
         self::assertEquals($today, $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
         self::assertEquals(new DateInterval('P0D'), $period->getDateInterval());
     }
 
@@ -479,8 +472,6 @@ class FunctionTest extends TestCase
         self::assertEquals(new DateTimeImmutable('2008-08-01T00:00:00+08:00'), $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
     }
 
     public function testYearWithDateTimeInterface(): void
@@ -491,7 +482,5 @@ class FunctionTest extends TestCase
         self::assertEquals(new DateTimeImmutable('2009-01-01T00:00:00+08:00'), $period->getEndDate());
         self::assertEquals('+08:00', $period->getStartDate()->format('P'));
         self::assertEquals('+08:00', $period->getEndDate()->format('P'));
-        self::assertInstanceOf(ExtendedDate::class, $period->getStartDate());
-        self::assertInstanceOf(ExtendedDate::class, $period->getEndDate());
     }
 }